feat: scoped teacher permissions by class/subject + tiered exam permissions + admin-only settings
- Replace broad academics.manage with academics.settings.manage (admin-only)
- Add TeacherScopeService + TeacherScoped trait for query scoping
- 10 new granular permissions for exam basic/elevated tiers
- 22+ controllers updated with new middleware + scoping
- Gate::before in AuthServiceProvider for runtime resolution
- Update sidebar menu with correct permission gating
- RbacVisibilityTest: 14 hermetic tests for 3 roles + scoping
- Fix TermController middleware and sidebar academic permissions

+++ database/migrations/2026_02_04_063502_add_department_id_to_subjects_table.php
@@ -13,7 +13,8 @@ return new class extends Migration
     {
         Schema::table('subjects', function (Blueprint $table) {
             if (!Schema::hasColumn('subjects', 'department_id')) {
-                $table->unsignedBigInteger('department_id')->nullable()->after('subject_id');
+                // Must match departments.department_id (signed INT) or FK errno 150
+                $table->integer('department_id')->nullable()->after('subject_id');
                 $table->foreign('department_id')->references('department_id')->on('departments')->onDelete('set null');
             }
         });
